Update OperatingSystemsFunctions.txt
I didn't go into detail on the drivers but they are device controllers that are for each device type to limit the work put on the operating system like a printer would have its own driver called a printer controller. this would help to prevent confusion for us and also keep a better watch on each device.